Move SimulationInputs to settings.py
